+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0] - 2025-07-24
+- Feat: `x-venue` in websocket subscriptions
+
+## [2.4.0] - 2025-07-08
+
+- Feat: send alert message
+
 ## [2.3.2] - 2025-01-28
 
 - Fix: payment 0